# Always send recording_id as part of LogMsg