--- a/rerun_py/src/arrow.rs
+++ b/rerun_py/src/arrow.rs
@@ -9,9 +9,7 @@ use pyo3::{
     types::{IntoPyDict, PyString},
     PyAny, PyResult,
 };
-use re_log_types::{
-    component_types, DataCell, DataRow, DataTableError, EntityPath, LogMsg, MsgId, TimePoint,
-};
+use re_log_types::{component_types, DataCell, DataRow, DataTable, EntityPath, MsgId, TimePoint};
 
 /// Perform conversion between a pyarrow array to arrow2 types.
 ///
@@ -88,7 +86,7 @@ pub fn build_chunk_from_components(
     entity_path: &EntityPath,
     components: &PyDict,
     time_point: &TimePoint,
-) -> PyResult<LogMsg> {
+) -> PyResult<DataTable> {
     let (arrays, fields): (Vec<Box<dyn Array>>, Vec<Field>) = itertools::process_results(
         components.iter().map(|(name, array)| {
             let name = name.downcast::<PyString>()?.to_str()?;
@@ -112,9 +110,7 @@ pub fn build_chunk_from_components(
         cells,
     );
 
-    let msg = (&row.into_table())
-        .try_into()
-        .map_err(|err: DataTableError| PyValueError::new_err(err.to_string()))?;
+    let data_table = row.into_table();
 
-    Ok(LogMsg::ArrowMsg(msg))
+    Ok(data_table)
 }