… crates/re_data_store/examples/memory_usage.rs